requirements.txt
# Core Vision
opencv-python>=4.8.0
mediapipe>=0.10.0
facenet-pytorch>=2.5.3

# Object Detection
ultralytics>=8.0.0

# Machine Learning Framework
torch>=1.7.0
torchvision>=0.8.0
torchaudio>=2.1.0

# Audio Processing
pyaudio>=0.2.14
openai-whisper>=20250625

# Screen Capture
mss>=6.1.0

# Alert System
gTTS>=2.5.4
pygame>=2.6.1

# Reporting System
pdfkit>=1.0.0
jinja2>=3.0.0
matplotlib>=3.5.0
Pillow>=9.0.0

# System Utilities
psutil>=5.8.0
pyyaml>=5.0.0
numpy>=1.20.0
flask>=2.0.0

# Platform-Specific Dependencies
pywin32>=300; sys_platform == 'win32'
pygrabber>=0.2; sys_platform == 'win32'
dbus-python>=1.3.2; sys_platform == 'linux'
pygobject>=3.42.2; sys_platform == 'linux'